Extract error overlay into its own file
